backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
requests==2.32.3

# Google Cloud clients
google-cloud-storage==2.18.2
google-cloud-firestore==2.16.0
google-cloud-vision==3.8.0
google-cloud-tasks==2.19.3
google-auth==2.33.0

# PDF handling
pypdf==4.3.1

# Retry/backoff
tenacity==8.5.0
